+++ modules/chain-grabber/src/main/scala/org/ergoplatform/explorer/indexer/Application.scala
@@ -1,4 +1,4 @@
-package org.ergoplatform.explorer.grabber
+package org.ergoplatform.explorer.indexer
 
 import cats.effect.{ExitCode, Resource}
 import cats.syntax.functor._
@@ -7,7 +7,7 @@ import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
 import monix.eval.{Task, TaskApp}
 import org.ergoplatform.explorer.clients.ergo.ErgoNetworkClient
 import org.ergoplatform.explorer.db.DoobieTrans
-import org.ergoplatform.explorer.grabber.processes.NetworkViewSync
+import org.ergoplatform.explorer.indexer.processes.ChainIndexing
 import org.ergoplatform.explorer.settings.GrabberAppSettings
 import org.http4s.client.blaze.BlazeClientBuilder
 
@@ -20,12 +20,12 @@ object Application extends TaskApp {
   def run(args: List[String]): Task[ExitCode] =
     resources(args.headOption).use {
       case (logger, settings, client, xa) =>
-        logger.info("Starting Grabber service ..") >>
+        logger.info("Starting Indexer service ..") >>
         ErgoNetworkClient[Task](client, settings.masterNodesAddresses).flatMap { ns =>
-          NetworkViewSync[Task, ConnectionIO](settings, ns)(xa)
+          ChainIndexing[Task, ConnectionIO](settings, ns)(xa)
             .flatMap(_.run.compile.drain)
             .as(ExitCode.Success)
-        }.guarantee(logger.info("Stopping Grabber service .."))
+        }.guarantee(logger.info("Stopping Indexer service .."))
     }
 
   private def resources(configPathOpt: Option[String]) =
@@ -33,6 +33,6 @@ object Application extends TaskApp {
       logger   <- Resource.liftF(Slf4jLogger.create)
       settings <- Resource.liftF(GrabberAppSettings.load(configPathOpt))
       client   <- BlazeClientBuilder[Task](global).resource
-      xa       <- DoobieTrans[Task]("GrabberPool", settings.db).map(_.trans)
+      xa       <- DoobieTrans[Task]("IndexerPool", settings.db).map(_.trans)
     } yield (logger, settings, client, xa)
 }
