+++ modules/chain-grabber/src/main/scala/org/ergoplatform/explorer/indexer/extractors/BlockInfoBuildFrom.scala
@@ -1,9 +1,9 @@
-package org.ergoplatform.explorer.grabber.extractors
+package org.ergoplatform.explorer.indexer.extractors
 
 import cats.Monad
 import org.ergoplatform.explorer.Err.{ProcessingErr, RefinementFailed}
 import org.ergoplatform.explorer.db.models.BlockInfo
-import org.ergoplatform.explorer.grabber.models.SlotData
+import org.ergoplatform.explorer.indexer.models.SlotData
 import org.ergoplatform.explorer.protocol.constants
 import org.ergoplatform.explorer.protocol.models.ApiFullBlock
 import org.ergoplatform.explorer.settings.ProtocolSettings
