+++ modules/chain-grabber/src/main/scala/org/ergoplatform/explorer/indexer/models/FlatBlock.scala
@@ -1,4 +1,4 @@
-package org.ergoplatform.explorer.grabber.models
+package org.ergoplatform.explorer.indexer.models
 
 import org.ergoplatform.explorer.db.models._
