+++ modules/chain-grabber/src/main/scala/org/ergoplatform/explorer/indexer/models/SlotData.scala
@@ -1,4 +1,4 @@
-package org.ergoplatform.explorer.grabber.models
+package org.ergoplatform.explorer.indexer.models
 
 import org.ergoplatform.explorer.db.models.BlockInfo
 import org.ergoplatform.explorer.protocol.models.ApiFullBlock
